Fix bucell cell filter generation
When creating burnup cell tallies to export to xml the CellFilter is now created only once and passed on to the various tally creation functions. While the previous way of creating it in each function is fine effect-wise it caused OpenMC to give out unnecessary, confusing warnings since the FilterID was reused.