Added option to redirect ONIX output
The user can now provide the burn routine with a prefix - relative to the current working directory - that the ONIX output is redirected to. The structure of the output itself remains unchanged. Note that ONIX still works and temporarily creates and reads files in the current directory, and as such running several instances of ONIX at the same time from the same directory will still cause issues, regardless of their final output directories being different.