Custom temperature treatment settings
ONIX now copies and applies user-specified settings for temperature treatment. If no method is given it defaults to 'nearest'. If the method is 'nearest' and no tolerance is given the latter defaults to 500 - if the method is 'interpolation' no default for tolerance is given (since it is not applied with this method). All other possible settings are purely optional and default to nothing.