Overhaul of plot_bucell_nuclide_network utility
Made some improvements to the plot_bucell_nuclide_network utility function:
Threshold can now be given as fraction of the total reaction rate instead of absolute reaction rate (the latter is still the default). The nodes in the created plot now form a proper circle, and the plot is created quadratic to avoid distortion of dimensions. Readibility of text has been improved. The plot is now saved for further use. Deleted some useless code snippets.